pii-frontend/pyproject.toml
[project]
name = "pii-frontend"
version = "0.1.0"
description = ""
authors = [
    {name = "Than Htut Soe",email = "than@consigli.no"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio (>=5.12.0,<6.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
